requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.0.1
django-allauth==0.56.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.1
sqlparse==0.2.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
